refactor(settings): extract WeightPage
