+++ tests/test_settings_window_weight_quick_add.py
@@ -1,23 +1,24 @@
-"""Regression/behavior test for the Weight tab's Quick Add panel
-(gui.settings_window_pyside.SettingsWindow._weight_quick_add_product).
+"""Regression/behavior test for the Weight page's Quick Add panel
+(gui.settings.weight.WeightPage._weight_quick_add_product).
 
-Exercises the bound methods directly against a bare QWidget standing in for
-SettingsWindow, rather than constructing the real SettingsWindow -- its full
-__init__ builds every settings tab and hangs under the offscreen QPA
-platform (no way to dismiss real dialogs headlessly), which this test has
-no need to go through just to check the Quick Add row-insertion logic.
+Exercises the bound methods directly against a WeightPage built via __new__
+(only QWidget.__init__ run, not WeightPage.__init__, which builds the full
+sub-tab UI), carrying just the attributes the Quick Add methods touch --
+that full build needs a weight_config/column_mappings/delimiter fixture this
+test has no need to go through just to check the Quick Add row-insertion
+logic.
 """
 import pytest
 from PySide6.QtWidgets import (
     QApplication,
     QCheckBox,
-    QDialog,
     QDoubleSpinBox,
     QLineEdit,
     QTableWidget,
+    QWidget,
 )
 
-from gui.settings_window_pyside import SettingsWindow
+from gui.settings.weight import WeightPage
 
 
 @pytest.fixture(scope="module")
@@ -27,13 +28,12 @@ def qapp():
 
 @pytest.fixture
 def sw(qapp):
-    """A real SettingsWindow instance with only QDialog.__init__ run (not
-    SettingsWindow.__init__, which builds every settings tab and hangs under
-    the offscreen QPA platform), carrying just the attributes the Quick Add
+    """A real WeightPage instance with only QWidget.__init__ run (not
+    WeightPage.__init__), carrying just the attributes the Quick Add
     methods touch. Using the real class (not a bare QWidget) keeps `self.` calls
     inside those methods resolving to the real bound methods."""
-    obj = SettingsWindow.__new__(SettingsWindow)
-    QDialog.__init__(obj)
+    obj = WeightPage.__new__(WeightPage)
+    QWidget.__init__(obj)
     obj.weight_products_table = QTableWidget(0, 7)
     obj.weight_divisor_spin = QDoubleSpinBox()
     obj.weight_divisor_spin.setValue(6000)
@@ -54,7 +54,7 @@ def test_quick_add_appends_a_row_and_resets_the_form(sw):
     sw.weight_quick_h.setValue(7)
     sw.weight_quick_no_pkg.setChecked(True)
 
-    SettingsWindow._weight_quick_add_product(sw)
+    WeightPage._weight_quick_add_product(sw)
 
     assert sw.weight_products_table.rowCount() == 1
     assert sw.weight_products_table.item(0, 0).text() == "NEWSKU"
@@ -67,7 +67,7 @@ def test_quick_add_appends_a_row_and_resets_the_form(sw):
 
 
 def test_quick_add_rejects_blank_sku(sw):
-    SettingsWindow._weight_quick_add_product(sw)
+    WeightPage._weight_quick_add_product(sw)
     assert sw.weight_products_table.rowCount() == 0
 
 
@@ -76,9 +76,9 @@ def test_quick_add_rejects_duplicate_sku_without_adding_a_row(sw, monkeypatch):
     monkeypatch.setattr(QMessageBox, "warning", lambda *a, **k: None)
 
     sw.weight_quick_sku.setText("DUPE")
-    SettingsWindow._weight_quick_add_product(sw)
+    WeightPage._weight_quick_add_product(sw)
     assert sw.weight_products_table.rowCount() == 1
 
     sw.weight_quick_sku.setText("DUPE")
-    SettingsWindow._weight_quick_add_product(sw)
+    WeightPage._weight_quick_add_product(sw)
     assert sw.weight_products_table.rowCount() == 1
